cleanup install requirements
