+++ cert_issuer/bitcoin/connectors.py
@@ -18,9 +18,7 @@ from pycoin.services.providers import service_provider_methods
 from pycoin.tx.Spendable import Spendable
 
 from cert_issuer import helpers
-from cert_issuer.errors import ConnectorError, BroadcastError
-
-BROADCAST_RETRY_INTERVAL = 30
+from cert_issuer.errors import BroadcastError
 
 try:
     from urllib2 import urlopen, HTTPError
@@ -29,18 +27,8 @@ except ImportError:
     from urllib.request import urlopen, HTTPError
     from urllib.parse import urlencode
 
+BROADCAST_RETRY_INTERVAL = 30
 MAX_BROADCAST_ATTEMPTS = 3
-
-
-def try_get(url):
-    """throw error if call fails"""
-    response = requests.get(url)
-    if int(response.status_code) != 200:
-        error_message = 'Error! status_code={}, error={}'.format(
-            response.status_code, response.json()['error'])
-        logging.error(error_message)
-        raise ConnectorError(error_message)
-    return response
 
 
 def to_hex(transaction):
@@ -49,62 +37,7 @@ def to_hex(transaction):
     tx_as_hex = b2h(s.getvalue())
     return tx_as_hex
 
-class EtherscanBroadcaster(object):
-    def __init__(self, base_url):
-        self.base_url = base_url
-    
-    def broadcast_tx(self, tx, api_token):
-        tx_hex = tx
 
-        broadcast_url = self.base_url + '?module=proxy&action=eth_sendRawTransaction'
-        if api_token:
-            '&apikey=%s' % api_token
-        response = requests.post(broadcast_url, data={'hex': tx_hex})
-        if int(response.status_code) == 200:
-            tx_id = response.json().get('result', None)
-            logging.info("Transaction ID obtained from broadcast through Etherscan: %s", tx_id)
-            return tx_id
-        logging.error('Error broadcasting the transaction through the Etherscan API. Error msg: %s', response.text)
-        raise BroadcastError(response.text)
-    
-    def get_balance(self, address, api_token):
-        """
-        returns the balance in wei
-        with some inspiration from PyWallet
-        """
-        broadcast_url = self.base_url + '?module=account&action=balance' 
-        broadcast_url += '&address=%s' % address
-        broadcast_url += '&tag=latest'
-        if api_token:
-            '&apikey=%s' % api_token
-        response = requests.get(broadcast_url)
-        if int(response.status_code) == 200:
-            balance = int(response.json().get('result', None))
-            logging.info('Balance check succeeded: %s', response.json())
-            return balance
-        raise BroadcastError(response.text)
-    
-    def get_address_nonce(self, address, api_token):
-        """
-        Looks up the address nonce of this address
-        Neccesary for the transaction creation
-        """
-        broadcast_url = self.base_url + '?module=proxy&action=eth_getTransactionCount'
-        broadcast_url += '&address=%s' % address
-        broadcast_url += '&tag=latest'
-        if api_token:
-            '&apikey=%s' % api_token
-        response = requests.get(broadcast_url, )
-        if int(response.status_code) == 200:
-            #the int(res, 0) transforms the hex nonce to int
-            nonce = int(response.json().get('result', None), 0)
-            logging.info('Nonce check went correct: %s', response.json())
-            return nonce
-        else:
-            logging.info('response error checking nonce')
-        raise BroadcastError('Error checking the nonce through the Etherscan API. Error msg: %s', response.text)
-        
-        
 class BlockExplorerBroadcaster(object):
     def __init__(self, base_url):
         self.base_url = base_url
@@ -194,50 +127,6 @@ class ServiceProviderConnector(object):
     def broadcast_tx(self, tx):
         pass
 
-class EthereumServiceProviderConnector(ServiceProviderConnector):
-    #param local_node indicates if a local node is running or if the tx should be broadcast to external providers 
-    def __init__(self, ethereum_chain, api_key, local_node=False):
-        self.ethereum_chain = ethereum_chain
-        self.api_key = api_key
-        self.local_node = local_node
-    
-    def get_balance(self, address):
-        for m in get_providers_for_chain(self.ethereum_chain, self.local_node):
-            if isinstance(m, BitcoindConnector):
-                pass
-            else:    
-                try:
-                    logging.debug('m=%s', m)
-                    balance = m.get_balance(address, self.api_key)
-                    return balance
-                except Exception as e:
-                    logging.warning(e)
-                    pass            
-        return 0
-
-    def get_address_nonce(self, address):
-        for m in get_providers_for_chain(self.ethereum_chain, self.local_node):
-            try:
-                logging.debug('m=%s', m)
-                nonce = m.get_address_nonce(address, self.api_key)
-                return nonce
-            except Exception as e:
-                logging.warning(e)
-                pass
-        return 0
-
-    def broadcast_tx(self, tx):
-        for m in get_providers_for_chain(self.ethereum_chain, self.local_node):
-            try:
-                logging.debug('m=%s', m)
-                txid = m.broadcast_tx(tx, self.api_key)
-                return txid
-            except Exception as e:
-                logging.warning(e)
-                pass
-        ##in case of failure:
-        return '0xfail'
-         
 
 class BitcoinServiceProviderConnector(ServiceProviderConnector):
     def __init__(self, bitcoin_chain, bitcoind=False):
@@ -245,7 +134,8 @@ class BitcoinServiceProviderConnector(ServiceProviderConnector):
         self.bitcoind = bitcoind
 
     def spendables_for_address(self, bitcoin_address):
-        for m in service_provider_methods('spendables_for_address', get_providers_for_chain(self.bitcoin_chain, self.bitcoind)):
+        for m in service_provider_methods('spendables_for_address',
+                                          get_providers_for_chain(self.bitcoin_chain, self.bitcoind)):
             try:
                 logging.debug('m=%s', m)
                 spendables = m(bitcoin_address)
@@ -339,32 +229,22 @@ PYCOIN_XTN_PROVIDERS = "blockexplorer.com"  # chain.so
 # initialize connectors
 connectors = {}
 
-
 # configure mainnet providers
-provider_list = providers.providers_for_config_string(PYCOIN_BTC_PROVIDERS, helpers.to_pycoin_chain(Chain.bitcoin_mainnet))
+provider_list = providers.providers_for_config_string(PYCOIN_BTC_PROVIDERS,
+                                                      helpers.to_pycoin_chain(Chain.bitcoin_mainnet))
 provider_list.append(BlockrIOBroadcaster('https://btc.blockr.io/api/v1'))
 provider_list.append(BlockExplorerBroadcaster('https://blockexplorer.com/api'))
 provider_list.append(InsightProvider(netcode=helpers.to_pycoin_chain(Chain.bitcoin_mainnet)))
 provider_list.append(ChainSoProvider(netcode=helpers.to_pycoin_chain(Chain.bitcoin_mainnet)))
 connectors[Chain.bitcoin_mainnet] = provider_list
 
-
 # configure testnet providers
-xtn_provider_list = providers.providers_for_config_string(PYCOIN_XTN_PROVIDERS, helpers.to_pycoin_chain(Chain.bitcoin_testnet))
+xtn_provider_list = providers.providers_for_config_string(PYCOIN_XTN_PROVIDERS,
+                                                          helpers.to_pycoin_chain(Chain.bitcoin_testnet))
 xtn_provider_list.append(ChainSoProvider(netcode=helpers.to_pycoin_chain(Chain.bitcoin_testnet)))
 xtn_provider_list.append(BlockrIOBroadcaster('https://tbtc.blockr.io/api/v1'))
 xtn_provider_list.append(BlockExplorerBroadcaster('https://testnet.blockexplorer.com/api'))
 connectors[Chain.bitcoin_testnet] = xtn_provider_list
-
-#Configure Ethereum mainnet connectors 
-eth_provider_list = []
-eth_provider_list.append(EtherscanBroadcaster('https://api.etherscan.io/api'))
-connectors[Chain.ethereum_mainnet] = eth_provider_list
-
-#Configure Ethereum Ropsten testnet connectors
-rop_provider_list = []
-rop_provider_list.append(EtherscanBroadcaster('https://ropsten.etherscan.io/api'))
-connectors[Chain.ethereum_ropsten] = rop_provider_list
 
 
 def get_providers_for_chain(chain, bitcoind=False):
@@ -372,4 +252,3 @@ def get_providers_for_chain(chain, bitcoind=False):
         return [BitcoindConnector(helpers.to_pycoin_chain(chain))]
     else:
         return connectors[chain]
-
