+++ cert_issuer/bitcoin/tx_utils.py
@@ -16,50 +16,10 @@ from pycoin.tx.Tx import Tx, TxOut
 
 from cert_issuer.errors import UnverifiedTransactionError
 
-
-
 BYTES_PER_INPUT = 148  # assuming compressed public key
 BYTES_PER_OUTPUT = 34
 FIXED_EXTRA_BYTES = 10
 OP_RETURN_BYTE_COUNT = 43  # our op_return output values always have the same length because they are SHA-256 hashes
-
-COIN = 100000000  # satoshis in 1 btc
-
-
-class EthereumTransactionCostConstants(object):
-    def __init__(self, recommended_gas_price=20000000000, recommended_gas_limit=25000):
-        self.recommended_gas_price = recommended_gas_price
-        self.recommended_gas_limit = recommended_gas_limit
-        logging.info('Set cost constants to recommended_gas_price=%f, recommended_gas_limit=%f',
-                     self.recommended_gas_price, self.recommended_gas_limit)
-        
-    """
-    The below methods currently only use the supplied gasprice/limit.
-    These values can also be better estimated via a call to the Ethereum blockchain.
-    """
-    def get_recommended_max_cost(self):
-        return self.recommended_gas_price * self.recommended_gas_limit
-    
-    def get_gas_price(self):
-        return self.recommended_gas_price
-
-    def get_gas_limit(self):
-        return self.recommended_gas_limit
-
-
-class BitcoinTransactionCostConstants(object):
-    def __init__(self, recommended_tx_fee=0.0006, min_per_output=0.0000275, satoshi_per_byte=250):
-        self.recommended_tx_fee = recommended_tx_fee
-        self.min_per_output = min_per_output
-        self.satoshi_per_byte = satoshi_per_byte
-        logging.info('Set cost constants to recommended_tx_fee=%f,min_per_output=%f,satoshi_per_byte=%d',
-                     self.recommended_tx_fee, self.min_per_output, self.satoshi_per_byte)
-
-    def get_minimum_output_coin(self):
-        return self.min_per_output * COIN
-
-    def get_recommended_fee_coin(self):
-        return self.recommended_tx_fee * COIN
 
 
 def create_trx(op_return_val, issuing_transaction_fee, issuing_address, tx_outs, tx_inputs):
@@ -191,30 +151,3 @@ def calculate_tx_fee(tx_cost_constants, num_inputs, num_outputs):
     tx_size = calculate_raw_tx_size_with_op_return(num_inputs, num_outputs)
     tx_fee = tx_cost_constants.satoshi_per_byte * tx_size
     return max(tx_fee, tx_cost_constants.get_recommended_fee_coin())
-
-
-def create_Ethereum_trx(issuing_address, nonce, to_address, blockchain_bytes, gasprice, gaslimit):
-    #the actual value transfer is 0 in the Ethereum implementation
-    from ethereum.transactions import Transaction
-    value = 0
-    tx = Transaction(nonce=nonce, gasprice=gasprice, startgas=gaslimit, to=to_address, value=value, data=blockchain_bytes)
-    return tx
-
-def verify_eth_transaction(signed_hextx, ethDataField):
-    """
-    Verify ethDataField field in transaction
-    :param signed_hextx:
-    :param ethDataField:
-    :return:
-    """
-    logging.info('verifying ethDataField value for transaction')
-    ethdata_hash = []
-    for s in signed_hextx.split('80a0'):
-        ethdata_hash.append(s)
-    ethdata_hash = ethdata_hash[1][:64]
-    result = (ethDataField == ethdata_hash)
-    if not result:
-        error_message = 'There was a problem verifying the transaction'
-        raise UnverifiedTransactionError(error_message)
-    logging.info('verified ethDataField')
-
