+++ cert_issuer/models.py
@@ -1,0 +1,95 @@
+from abc import abstractmethod
+
+from cert_issuer.config import ESTIMATE_NUM_INPUTS
+
+
+class CertificateHandler(object):
+    @abstractmethod
+    def validate_certificate(self, certificate_metadata):
+        pass
+
+    @abstractmethod
+    def sign_certificate(self, signer, certificate_metadata):
+        pass
+
+    @abstractmethod
+    def get_byte_array_to_issue(self, certificate_metadata):
+        pass
+
+    @abstractmethod
+    def add_proof(self, certificate_metadata, merkle_proof):
+        pass
+
+
+class ServiceProviderConnector(object):
+    @abstractmethod
+    def get_balance(self, address):
+        pass
+
+    def broadcast_tx(self, tx):
+        pass
+
+
+class Signer(object):
+    """
+    Abstraction for a component that can sign.
+    """
+
+    def __init__(self):
+        pass
+
+    @abstractmethod
+    def sign_message(self, wif, message_to_sign):
+        pass
+
+    @abstractmethod
+    def sign_transaction(self, wif, transaction_to_sign):
+        pass
+
+
+class SecretManager(object):
+    def __init__(self, signer):
+        self.signer = signer
+        self.wif = None
+
+    @abstractmethod
+    def start(self):
+        pass
+
+    @abstractmethod
+    def stop(self):
+        pass
+
+    def sign_message(self, message_to_sign):
+        return self.signer.sign_message(self.wif, message_to_sign)
+
+    def sign_transaction(self, transaction_to_sign):
+        return self.signer.sign_transaction(self.wif, transaction_to_sign)
+
+
+class TransactionHandler(object):
+    @abstractmethod
+    def ensure_balance(self):
+        pass
+
+    @abstractmethod
+    def issue_transaction(self, blockchain_bytes):
+        pass
+
+
+class MockTransactionHandler(TransactionHandler):
+    def ensure_balance(self):
+        pass
+
+    def issue_transaction(self, op_return_bytes):
+        return 'This has not been issued on a blockchain and is for testing only'
+
+
+class TransactionCreator(object):
+    @abstractmethod
+    def estimate_cost_for_certificate_batch(self, tx_cost_constants, num_inputs=ESTIMATE_NUM_INPUTS):
+        pass
+
+    @abstractmethod
+    def create_transaction(self, tx_cost_constants, issuing_address, inputs, op_return_value):
+        pass
