+++ ethereum_requirements.txt
@@ -1,0 +1,3 @@
+-r requirements.txt
+ethereum>=2.1.0
+rlp>=0.6.0
