+++ requirements.txt
@@ -2,13 +2,9 @@ cert-core>=2.1.6
 cert-schema>=2.1.4
 chainpoint==0.0.2
 configargparse==0.12.0
-ethereum>=2.1.0
 glob2==0.6
 mock==2.0.0
-pycoin>=0.80
 requests[security]>=2.18.4
+pycoin>=0.80
 pyld>=0.8.1
-pysha3>=1.0.1
-python-bitcoinlib>=0.8.0
 tox>=2.7.0
-rlp>=0.6.0
